requirements.txt
psycopg2
bleach
django==3.1.8
django-composite-field
django-ordered-model
git+https://github.com/vporton/django-language-field.git
django-bootstrap3
django-recaptcha2
django-googlytics==0.2.1
django-defender
django-cities
